requirements.txt
bcrypt==3.1.7
bottle==0.12.18
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
dnspython==1.16.0
eventlet==0.25.2
greenlet==0.4.16
idna==2.9
monotonic==1.5
Naked==0.1.31
pycparser==2.20
pycryptodome==3.9.7
PyYAML==5.3.1
requests==2.23.0
shellescape==3.8.1
six==1.15.0
urllib3==1.25.9
